fix papi and extract demo
