+++ packages/registry/app/components/page.tsx
@@ -1,9 +1,9 @@
-import { AddressInputWithProvider } from "@/registry/dot-ui/blocks/address-input/components/address-input.papi";
+import AddressInputPage from "@/registry/dot-ui/blocks/address-input/demo/page.papi";
 
 export default function Docs() {
   return (
     <div className="flex flex-col gap-4 max-w-md mx-auto p-4">
-      <AddressInputWithProvider />
+      <AddressInputPage />
     </div>
   );
 }
