+++ packages/registry/registry.json
@@ -1,12 +1,12 @@
 {
   "$schema": "https://ui.shadcn.com/schema/registry.json",
-  "name": "dot-ui-papi",
+  "name": "dot-ui-dedot",
   "homepage": "https://dot-ui.com",
   "items": [
     {
       "name": "address-input",
       "type": "registry:component",
-      "title": "Address Input (papi)",
+      "title": "Address Input (dedot)",
       "description": "A comprehensive address input component with SS58/Ethereum validation, identity lookup using polkadot-api, and customizable identicon themes (polkadot, substrate, beachball, jdenticon).",
       "dependencies": [
         "@polkadot/keyring",
